working on debugging the gpu program

+++ gpu-parallel/test/bcsstk13_test.cu
@@ -10,10 +10,10 @@
 int main()
 {
     // BCSSTK13
-    double* cg_solution;
-    double* FFGE_matrix;
-    double b_13[2003];
-    double init_13[2003];
+    float* cg_solution;
+    float* FFGE_matrix;
+    float b_13[2003];
+    float init_13[2003];
     float startTime;
     float endTime;
     float timeElapsed;
